refactor(idpe-17789): rename partition_stream mod to compaction_job_stream mod
