+++ compactor/src/components/mod.rs
@@ -2,17 +2,18 @@ use std::sync::Arc;
 
 use self::{
     changed_files_filter::ChangedFilesFilter, commit::CommitToScheduler,
-    df_plan_exec::DataFusionPlanExec, df_planner::DataFusionPlanner, divide_initial::DivideInitial,
-    file_classifier::FileClassifier, ir_planner::IRPlanner, parquet_files_sink::ParquetFilesSink,
+    compaction_job_stream::CompactionJobStream, df_plan_exec::DataFusionPlanExec,
+    df_planner::DataFusionPlanner, divide_initial::DivideInitial, file_classifier::FileClassifier,
+    ir_planner::IRPlanner, parquet_files_sink::ParquetFilesSink,
     partition_done_sink::PartitionDoneSink, partition_files_source::PartitionFilesSource,
     partition_filter::PartitionFilter, partition_info_source::PartitionInfoSource,
-    partition_stream::CompactionJobStream,
     post_classification_partition_filter::PostClassificationPartitionFilter,
     round_info_source::RoundInfoSource, round_split::RoundSplit, scratchpad::ScratchpadGen,
 };
 
 pub mod changed_files_filter;
 pub(crate) mod commit;
+pub mod compaction_job_stream;
 pub mod compaction_jobs_source;
 pub mod df_plan_exec;
 pub mod df_planner;
@@ -30,7 +31,6 @@ pub mod partition_files_source;
 pub mod partition_filter;
 pub mod partition_info_source;
 pub mod partition_source;
-pub mod partition_stream;
 pub mod post_classification_partition_filter;
 pub mod report;
 pub mod round_info_source;
